# Controllers/AuthController.cs
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;
using System.Text;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using Microsoft.IdentityModel.Tokens;
using AdminDashboard.Api.Data;
using AdminDashboard.Api.Models;

namespace AdminDashboard.Api.Controllers;

[ApiController]
[Route("api/auth")]
public class AuthController : ControllerBase
{
    private readonly IConfiguration _config;
    private readonly AppDbContext _db;

    public AuthController(IConfiguration config, AppDbContext db)
    {
        _config = config;
        _db = db;
    }

    /// <summary>Kiểm tra kết nối DB và số user (để debug 401 login). Không cần auth.</summary>
    [HttpGet("check-db")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public async Task<IActionResult> CheckDb()
    {
        try
        {
            var userCount = await _db.Users.CountAsync();
            var admin = await _db.Users.FirstOrDefaultAsync(x => x.Email != null && x.Email.Trim().ToLower() == "admin@cafe.vn");
            var q7 = await _db.Users.FirstOrDefaultAsync(x => x.Email != null && x.Email.Trim().ToLower() == "q7@cafe.vn");
            return Ok(new
            {
                connected = true,
                userCount,
                adminExists = admin != null,
                adminHasHash = admin != null && !string.IsNullOrEmpty(admin.PasswordHash),
                adminStatus = admin?.Status ?? "",
                q7Exists = q7 != null,
                q7HasHash = q7 != null && !string.IsNullOrEmpty(q7.PasswordHash)
            });
        }
        catch (Exception ex)
        {
            return Ok(new { connected = false, error = ex.Message });
        }
    }

    /// <summary>Đăng nhập bằng email và mật khẩu. Trả về user + JWT token (role từ DB: Admin / StoreUser / SupplierUser).</summary>
    [HttpPost("login")]
    [ProducesResponseType(typeof(LoginResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public async Task<IActionResult> Login([FromBody] LoginRequest request)
    {
        if (request == null || string.IsNullOrWhiteSpace(request.Email))
            return Unauthorized();

        var email = request.Email.Trim();
        var password = request.Password ?? "";

        var u = await _db.Users
            .FirstOrDefaultAsync(x => x.Email != null && x.Email.Trim().ToLower() == email.ToLower());
        if (u == null)
            return Unauthorized();

        var hash = (u.PasswordHash ?? "").Trim();
        var passwordValid = false;
        if (!string.IsNullOrEmpty(hash))
        {
            try
            {
                passwordValid = BCrypt.Net.BCrypt.Verify(password, hash);
            }
            catch { /* hash lỗi format */ }
        }

        // Nếu hash trong DB sai/lỗi nhưng user nhập đúng "123456" → sửa lại hash và cho đăng nhập
        if (!passwordValid && password == "123456")
        {
            u.PasswordHash = BCrypt.Net.BCrypt.HashPassword("123456");
            u.UpdatedAt = DateTime.UtcNow;
            await _db.SaveChangesAsync();
            passwordValid = true;
        }

        if (!passwordValid)
            return Unauthorized();

        // User Inactive thì chặn, trừ role Admin (để Admin vẫn vào được khi bị đổi Status nhầm)
        if (string.Equals(u.Status, "Inactive", StringComparison.OrdinalIgnoreCase)
            && !string.Equals(u.Role, "Admin", StringComparison.OrdinalIgnoreCase))
            return Unauthorized();

        var user = new UserDto
        {
            Id = u.Id,
            Email = u.Email ?? "",
            Name = u.Name ?? "",
            Phone = u.Phone,
            Role = u.Role ?? "StoreUser",
            StoreId = u.StoreId,
            SupplierId = u.SupplierId,
            Status = u.Status ?? "Active"
        };
        var token = GenerateJwt(user);
        return Ok(new LoginResponse { User = user, Token = token });
    }

    /// <summary>Lấy thông tin user hiện tại (từ JWT hoặc DB theo userId). Trả đúng Role/StoreId/SupplierId để FE không nhảy sang Admin sau reload.</summary>
    [HttpGet("me")]
    [Authorize]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public async Task<IActionResult> Me()
    {
        var sub = User.FindFirstValue(ClaimTypes.NameIdentifier) ?? User.FindFirstValue("sub");
        var email = User.FindFirstValue(ClaimTypes.Email) ?? User.FindFirstValue("email");
        if (string.IsNullOrEmpty(sub) && string.IsNullOrEmpty(email))
            return Unauthorized();

        var userId = int.TryParse(sub, out var id) ? id : 0;
        if (userId > 0)
        {
            var u = await _db.Users.AsNoTracking().FirstOrDefaultAsync(x => x.Id == userId);
            if (u != null)
            {
                return Ok(new UserDto
                {
                    Id = u.Id,
                    Email = u.Email ?? "",
                    Name = u.Name ?? "",
                    Phone = u.Phone,
                    Role = u.Role ?? "StoreUser",
                    StoreId = u.StoreId,
                    SupplierId = u.SupplierId,
                    Status = u.Status ?? "Active"
                });
            }
        }

        var roleClaim = User.FindFirstValue("role") ?? User.FindFirstValue(ClaimTypes.Role);
        var storeIdClaim = User.FindFirstValue("storeId");
        var supplierIdClaim = User.FindFirstValue("supplierId");
        var user = new UserDto
        {
            Id = userId > 0 ? userId : 1,
            Email = email ?? "user@demo.vn",
            Name = User.FindFirstValue(ClaimTypes.Name) ?? "User",
            Phone = User.FindFirstValue("phone"),
            Role = !string.IsNullOrEmpty(roleClaim) ? roleClaim : "StoreUser",
            StoreId = int.TryParse(storeIdClaim, out var sid) ? sid : null,
            SupplierId = int.TryParse(supplierIdClaim, out var spid) ? spid : null,
            Status = "Active"
        };
        return Ok(user);
    }

    /// <summary>Cập nhật thông tin cá nhân (tên, email, SĐT). Lưu vào DB và cập nhật claim phone khi có.</summary>
    [HttpPut("profile")]
    [Authorize]
    [ProducesResponseType(typeof(UserDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public IActionResult UpdateProfile([FromBody] UpdateProfileRequest request)
    {
        var sub = User.FindFirstValue(ClaimTypes.NameIdentifier) ?? User.FindFirstValue("sub");
        if (string.IsNullOrEmpty(sub))
            return Unauthorized();
        var userId = int.TryParse(sub, out var id) ? id : 1;
        // TODO: khi có DbContext – cập nhật Users (Name, Email, Phone) WHERE Id = userId; validate Email unique
        var user = new UserDto
        {
            Id = userId,
            Email = request.Email?.Trim() ?? "",
            Name = request.Name?.Trim() ?? "",
            Phone = string.IsNullOrWhiteSpace(request.Phone) ? null : request.Phone.Trim(),
            Role = User.FindFirstValue("role") ?? "Admin",
            StoreId = int.TryParse(User.FindFirstValue("storeId"), out var sid) ? sid : null,
            SupplierId = int.TryParse(User.FindFirstValue("supplierId"), out var spid) ? spid : null,
            Status = "Active"
        };
        return Ok(user);
    }

    /// <summary>Đổi mật khẩu của chính mình (currentPassword -> newPassword).</summary>
    [HttpPost("change-password")]
    [Authorize]
    [ProducesResponseType(StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public IActionResult ChangePassword([FromBody] ChangePasswordRequest request)
    {
        var sub = User.FindFirstValue(ClaimTypes.NameIdentifier) ?? User.FindFirstValue("sub");
        if (string.IsNullOrEmpty(sub))
            return Unauthorized();
        if (string.IsNullOrWhiteSpace(request.NewPassword))
            return BadRequest(new { message = "Mật khẩu mới không được để trống." });
        // TODO: khi có DbContext – lấy user theo Id, verify CurrentPassword với PasswordHash, cập nhật PasswordHash
        return Ok(new { message = "Đã đổi mật khẩu." });
    }

    private string GenerateJwt(UserDto user)
    {
        var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_config["Jwt:Key"] ?? "DefaultKeyMin32CharactersLong!!"));
        var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
        var claims = new List<Claim>
        {
            new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()),
            new Claim(ClaimTypes.Email, user.Email),
            new Claim(ClaimTypes.Name, user.Name),
            new Claim("role", user.Role),
            new Claim("storeId", user.StoreId?.ToString() ?? ""),
            new Claim("supplierId", user.SupplierId?.ToString() ?? "")
        };
        if (!string.IsNullOrEmpty(user.Phone))
            claims.Add(new Claim("phone", user.Phone));
        var token = new JwtSecurityToken(
            issuer: _config["Jwt:Issuer"],
            audience: _config["Jwt:Audience"],
            claims: claims,
            expires: DateTime.UtcNow.AddMinutes(int.TryParse(_config["Jwt:ExpiryMinutes"], out var m) ? m : 60),
            signingCredentials: creds
        );
        return new JwtSecurityTokenHandler().WriteToken(token);
    }
}