# openapi.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API cho hệ thống đặt hàng đa NCC (Multi-Supplier Order). Chi tiết xem API-FE-LIEN-KET.md
  version: 1.0.0
servers:
  - url: /api
    description: Base path API
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
  /auth/me:
    get:
      tags: [Auth]
      summary: Lấy thông tin user hiện tại
      security: [{ Bearer: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /categories:
    get:
      tags: [Categories]
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      tags: [Categories]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
  /categories/{id}:
    put:
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
    delete:
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token trả về từ POST /auth/login
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [Admin, StoreUser, SupplierUser] }
        storeId: { type: integer, nullable: true }
        supplierId: { type: integer, nullable: true }
        status: { type: string }
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
    CategoryInput:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
...